+++ Money.Data/Entities/Spending.cs
@@ -5,7 +5,7 @@
         public ulong Id { get; set; }
         public DateOnly Date { get; set; }
         public string Description { get; set; }
-        public decimal Ammount { get; set; }
+        public double Ammount { get; set; }
         public DateTime AddedOn { get; set; }
 
         public Spending()
