+++ Money.Data/IWriteOnlyData.cs
@@ -4,7 +4,7 @@ namespace Money.Data
 {
     public interface IWriteOnlyData
     {
-        ulong Insert(decimal ammount, string text, DateOnly date);
+        ulong Insert(double ammount, string text, DateOnly date);
         int Import(IEnumerable<SerializableSpending> toImport);
     }
 }
